created user auth api-views

+++ authentication/utils.py
@@ -14,9 +14,10 @@ def phone_number_validation(value):
 def is_user_created(request):
     from .models import User
     data = request.data
-    if ('phone_number' in data and
-            User.objects.filter(phone_number=data.get('phone_number')).exclude(id=request.user.id).exists()):
-        raise CustomApiException(ErrorCodes.ALREADY_EXISTS)
+    if 'phone_number' in data:
+        user = User.objects.filter(phone_number=data.get('phone_number')).exclude(id=request.user.id).first()
+        return user
+    return
 
 
 def gen_otp_code():
@@ -33,11 +34,18 @@ def check_block_user(user_id):
 
 def otp_create(user_id):
     from .models import OTP
+
     otp_count = OTP.objects.filter(user_id=user_id).count()
     if otp_count and otp_count > 2 and not check_block_user(user_id):
         raise CustomApiException(ErrorCodes.INVALID_INPUT, message='Too many attempts, please try again later.')
+
+    last_otp = OTP.objects.filter(user_id=user_id).first()
+    if last_otp and last_otp.created_at + timedelta(minutes=2) > datetime.now():
+        raise CustomApiException(ErrorCodes.INVALID_INPUT, message='OTP is not expired.')
+
     if otp_count and otp_count > 2 and check_block_user(user_id):
         OTP.objects.filter(user_id=user_id).delete()
+
     otp_code = gen_otp_code()
     otp = OTP.objects.create(user_id=user_id, otp_code=otp_code)
     send_otp_code(otp)
@@ -61,4 +69,5 @@ def otp_verification(data):
         otp.save(update_fields=['request_count'])
         raise CustomApiException(ErrorCodes.INVALID_INPUT, message='OTP code is invalid')
 
+    OTP.objects.filter(user_id=otp.user_id).delete()
     return otp.user_id
